moved Dockerfiles to within sub project directories
This project now has more of a monorepo feel to it, as all of the files
necessary to run each sub project are contained with the sub project
directories.

Notice that this commit also changes the hostnames the client uses to
send requests to the api service; they now use the service names given
in the top-level docker-compose.yml file. So this will break things if
someone tries to run each sub service individually (using docker run
...) but that should never happen anyway (devs should always use
docker-compose up) so it should be fine.